telegram-bot/python-aiogram-empty/pyproject.toml
[tool.poetry]
name = "example"
version = "0.1.0"
description = ""
authors = ["Sergey Fomchukov <durovam06@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
aiosqlite = "^0.21.0"
aiogram = "^3.17.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
